requirement.txt
ago==0.0.93
asgiref==3.3.4
attrs==21.2.0
Automat==20.2.0
awscli==1.19.78
beautifulsoup4==4.9.3
botocore==1.20.78
bs4==0.0.1
cchardet==2.1.7
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
chromedriver-binary==91.0.4472.19.0
click==8.0.1
colorama==0.4.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
cssselect==1.1.0
dateparser==1.0.0
Django==3.2.3
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
docutils==0.15.2
dotmap==1.3.23
drf-yasg==1.20.0
elasticsearch==7.12.1
fake-useragent==0.1.11
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
GoogleNews==1.5.8
h2==3.2.0
hjson==3.0.2
hpack==3.0.0
hurry.filesize==0.9
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==4.0.1
incremental==21.3.0
inflection==0.5.1
itemadapter==0.2.0
itemloaders==1.0.4
itypes==1.2.0
jieba3k==0.35.1
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
langdetect==1.0.9
lxml==4.6.3
MarkupSafe==2.0.1
news-fetch==0.2.8
news-please==1.5.21
newspaper3k==0.2.8
nltk==3.6.2
numpy==1.20.3
packaging==20.9
pandas==1.2.4
parsel==1.6.0
Pillow==8.2.0
plac==1.3.3
priority==1.3.0
Protego==0.1.16
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyJWT==2.1.0
PyMySQL==1.0.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
pywin32==300
PyYAML==5.4.1
queuelib==1.6.1
readability-lxml==0.8.1
regex==2021.4.4
requests==2.25.1
requests-file==1.5.1
rsa==4.7.2
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
s3transfer==0.4.2
Scrapy==2.5.0
selenium==3.141.0
service-identity==21.1.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tinysegmenter==0.3
tldextract==3.1.0
tqdm==4.60.0
Twisted==21.2.0
twisted-iocpsupport==1.0.1
typing-extensions==3.10.0.0
tzlocal==2.1
Unidecode==1.2.0
uritemplate==3.0.1
urllib3==1.26.4
w3lib==1.22.0
warcio==1.7.4
zipp==3.4.1
zope.interface==5.4.0
